requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
discord.py==1.7.3
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-core==1.7.2
google-cloud-translate==2.0.1
googleapis-common-protos==1.56.0
python-dotenv==0.20.0
requests==2.27.1
urllib3==1.26.9
Flask~=2.1.1